+++ NEWS.md
@@ -7,6 +7,7 @@
 * Add check to `gcs_get_bucket()` to only expect length 1 character vectors for bucket name. (#60)
 * Add paging for `gcs_object_list` (#58 - thanks @G3rtjan)
 * Add `saveToDisk` option to `gcs_load` (#52 - thanks @tomsing1)
+* Supply your own parse function to `gcs_get_object()` (#63 - thanks @nkeriks)
 
 # 0.2.0